pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["Fernando Gonzalez <contact@nandodev.nev>"]

[tool.poetry.dependencies]
python = "^3.10.2"
Django = "^4.2"
autopep8 = "^2.0.2"
flake8 = "^6.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
django-environ = "^0.9.0"
redis = "^4.3.4"
psycopg2-binary = "^2.9.3"
django-celery-beat = "^2.4.0"
django-timescaledb = "^0.2.12"
django-model-utils = "^4.3.1"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
